+++ sdk/appconfiguration/azure-appconfiguration-provider/azure/appconfiguration/provider/_azureappconfigurationprovider.py
@@ -119,6 +119,9 @@ def load(  # pylint: disable=docstring-keyword-should-match-keyword-only
     :keyword replica_discovery_enabled: Optional flag to enable or disable the discovery of replica endpoints. Default
      is True.
     :paramtype replica_discovery_enabled: bool
+    :keyword load_balancing_enabled: Optional flag to enable or disable the load balancing of replica endpoints. Default
+     is False.
+    :paramtype load_balancing_enabled: bool
     """
 
 
@@ -179,6 +182,9 @@ def load(  # pylint: disable=docstring-keyword-should-match-keyword-only
     :keyword replica_discovery_enabled: Optional flag to enable or disable the discovery of replica endpoints. Default
      is True.
     :paramtype replica_discovery_enabled: bool
+    :keyword load_balancing_enabled: Optional flag to enable or disable the load balancing of replica endpoints. Default
+     is False.
+    :paramtype load_balancing_enabled: bool
     """
 
 
@@ -227,17 +233,9 @@ def load(*args, **kwargs) -> "AzureAppConfigurationProvider":
     )
 
     provider = _buildprovider(connection_string, endpoint, credential, uses_key_vault=uses_key_vault, **kwargs)
-    headers = _get_headers(
-        kwargs.pop("headers", {}),
-        "Startup",
-        provider._replica_client_manager.get_client_count() - 1,  # pylint:disable=protected-access
-        provider._feature_flag_enabled,  # pylint:disable=protected-access
-        provider._feature_filter_usage,  # pylint:disable=protected-access
-        provider._uses_key_vault,  # pylint:disable=protected-access
-    )
 
     try:
-        provider._load_all(headers=headers)  # pylint:disable=protected-access
+        provider._load_all()  # pylint:disable=protected-access
     except Exception as e:
         _delay_failure(start_time)
         raise e
@@ -253,7 +251,16 @@ def _delay_failure(start_time: datetime.datetime) -> None:
         time.sleep((min_time - (current_time - start_time)).total_seconds())
 
 
-def _get_headers(headers, request_type, replica_count, uses_feature_flags, feature_filters_used, uses_key_vault) -> str:
+def _get_headers(
+    headers,
+    request_type,
+    replica_count,
+    uses_feature_flags,
+    feature_filters_used,
+    uses_key_vault,
+    uses_load_balancing,
+    failover_request,
+) -> Dict[str, str]:
     if os.environ.get(REQUEST_TRACING_DISABLED_ENVIRONMENT_VARIABLE, default="").lower() == "true":
         return headers
     correlation_context = "RequestType=" + request_type
@@ -290,6 +297,12 @@ def _get_headers(headers, request_type, replica_count, uses_feature_flags, featu
 
     if replica_count > 0:
         correlation_context += ",ReplicaCount=" + str(replica_count)
+
+    if uses_load_balancing:
+        correlation_context += ",UsesLoadBalancing"
+
+    if failover_request:
+        correlation_context += ",FailoverRequest"
 
     headers["Correlation-Context"] = correlation_context
     return headers
@@ -467,6 +480,8 @@ class AzureAppConfigurationProvider(Mapping[str, Union[str, JSON]]):  # pylint: 
         min_backoff: int = min(kwargs.pop("min_backoff", 30), interval)
         max_backoff: int = min(kwargs.pop("max_backoff", 600), interval)
 
+        self._uses_load_balancing = kwargs.pop("load_balancing_enabled", False)
+
         self._replica_client_manager = ConfigurationClientManager(
             connection_string=kwargs.pop("connection_string", None),
             endpoint=endpoint,
@@ -477,6 +492,7 @@ class AzureAppConfigurationProvider(Mapping[str, Union[str, JSON]]):  # pylint: 
             replica_discovery_enabled=kwargs.pop("replica_discovery_enabled", True),
             min_backoff_sec=min_backoff,
             max_backoff_sec=max_backoff,
+            load_balance=self._uses_load_balancing,
             **kwargs
         )
         self._dict: Dict[str, Any] = {}
@@ -510,7 +526,7 @@ class AzureAppConfigurationProvider(Mapping[str, Union[str, JSON]]):  # pylint: 
         self._update_lock = Lock()
         self._refresh_lock = Lock()
 
-    def refresh(self, **kwargs) -> None:
+    def refresh(self, **kwargs) -> None:  # pylint: disable=too-many-statements
         if not self._refresh_on and not self._feature_flag_refresh_enabled:
             logging.debug("Refresh called but no refresh enabled.")
             return
@@ -526,9 +542,10 @@ class AzureAppConfigurationProvider(Mapping[str, Union[str, JSON]]):  # pylint: 
                         Failed to refresh configuration settings from Azure App Configuration.
                         """
         exception: Exception = RuntimeError(error_message)
+        has_failed = False
         try:
             self._replica_client_manager.refresh_clients()
-            active_clients = self._replica_client_manager.get_active_clients()
+            self._replica_client_manager.find_active_clients()
 
             headers = _get_headers(
                 kwargs.pop("headers", {}),
@@ -537,8 +554,14 @@ class AzureAppConfigurationProvider(Mapping[str, Union[str, JSON]]):  # pylint: 
                 self._feature_flag_enabled,
                 self._feature_filter_usage,
                 self._uses_key_vault,
+                self._uses_load_balancing,
+                has_failed,
             )
-            for client in active_clients:
+
+            while self._replica_client_manager.has_next_client():
+                client = self._replica_client_manager.get_next_client()
+                if not client:
+                    return
                 try:
                     if self._refresh_on:
                         need_refresh, self._refresh_on, configuration_settings = client.refresh_configuration_settings(
@@ -556,11 +579,13 @@ class AzureAppConfigurationProvider(Mapping[str, Union[str, JSON]]):  # pylint: 
                         if need_refresh:
                             self._dict = configuration_settings_processed
                     if self._feature_flag_refresh_enabled:
-                        need_ff_refresh, self._refresh_on_feature_flags, feature_flags, filters_used = (
+                        need_ff_refresh, refresh_on_feature_flags, feature_flags, filters_used = (
                             client.refresh_feature_flags(
                                 self._refresh_on_feature_flags, self._feature_flag_selectors, headers, **kwargs
                             )
                         )
+                        if refresh_on_feature_flags:
+                            self._refresh_on_feature_flags = refresh_on_feature_flags
                         self._feature_filter_usage = filters_used
 
                         if need_refresh or need_ff_refresh:
@@ -573,7 +598,7 @@ class AzureAppConfigurationProvider(Mapping[str, Union[str, JSON]]):  # pylint: 
                 except AzureError as e:
                     exception = e
                     self._replica_client_manager.backoff(client)
-
+                    has_failed = True
             if not success:
                 self._refresh_timer.backoff()
                 if self._on_refresh_error:
@@ -586,9 +611,13 @@ class AzureAppConfigurationProvider(Mapping[str, Union[str, JSON]]):  # pylint: 
             self._refresh_lock.release()
 
     def _load_all(self, **kwargs):
-        active_clients = self._replica_client_manager.get_active_clients()
+        self._replica_client_manager.find_active_clients()
+        has_failed = False
 
-        for client in active_clients:
+        while self._replica_client_manager.has_next_client():
+            client = self._replica_client_manager.get_next_client()
+            if not client:
+                return
             try:
                 configuration_settings, sentinel_keys = client.load_configuration_settings(
                     self._selects, self._refresh_on, **kwargs
@@ -609,7 +638,16 @@ class AzureAppConfigurationProvider(Mapping[str, Union[str, JSON]]):  # pylint: 
                 for (key, label), etag in self._refresh_on.items():
                     if not etag:
                         try:
-                            headers = kwargs.get("headers", {})
+                            headers = _get_headers(
+                                kwargs.pop("headers", {}),
+                                "Startup",
+                                self._replica_client_manager.get_client_count() - 1,  # pylint:disable=protected-access
+                                self._feature_flag_enabled,  # pylint:disable=protected-access
+                                self._feature_filter_usage,  # pylint:disable=protected-access
+                                self._uses_key_vault,  # pylint:disable=protected-access
+                                self._uses_load_balancing,  # pylint:disable=protected-access
+                                has_failed,
+                            )
                             sentinel = client.get_configuration_setting(key, label, headers=headers)  # type:ignore
                             self._refresh_on[(key, label)] = sentinel.etag  # type:ignore
                         except HttpResponseError as e:
@@ -632,6 +670,7 @@ class AzureAppConfigurationProvider(Mapping[str, Union[str, JSON]]):  # pylint: 
                 return
             except AzureError:
                 self._replica_client_manager.backoff(client)
+                has_failed = True
         raise RuntimeError(
             "Failed to load configuration settings. No Azure App Configuration stores successfully loaded from."
         )