+++ sdk/appconfiguration/azure-appconfiguration-provider/azure/appconfiguration/provider/_client_manager.py
@@ -6,6 +6,7 @@
 from logging import getLogger
 import json
 import time
+import random
 from dataclasses import dataclass
 from typing import Tuple, Union, Dict, List, Any, Optional, Mapping
 from typing_extensions import Self
@@ -134,7 +135,7 @@ class _ConfigurationClientWrapper(_ConfigurationClientWrapperBase):
 
     @distributed_trace
     def load_configuration_settings(
-        self, selects: List[SettingSelector], refresh_on: Dict[Tuple[str, str], str], **kwargs
+        self, selects: List[SettingSelector], refresh_on: Mapping[Tuple[str, str], Optional[str]], **kwargs
     ) -> Tuple[List[ConfigurationSetting], Dict[Tuple[str, str], str]]:
         configuration_settings = []
         sentinel_keys = kwargs.pop("sentinel_keys", refresh_on)
@@ -186,8 +187,12 @@ class _ConfigurationClientWrapper(_ConfigurationClientWrapperBase):
 
     @distributed_trace
     def refresh_configuration_settings(
-        self, selects: List[SettingSelector], refresh_on: Dict[Tuple[str, str], str], headers: Dict[str, str], **kwargs
-    ) -> Tuple[bool, Dict[Tuple[str, str], str], List[Any]]:
+        self,
+        selects: List[SettingSelector],
+        refresh_on: Mapping[Tuple[str, str], Optional[str]],
+        headers: Dict[str, str],
+        **kwargs
+    ) -> Tuple[bool, Mapping[Tuple[str, str], Optional[str]], List[Any]]:
         """
         Gets the refreshed configuration settings if they have changed.
 
@@ -222,7 +227,7 @@ class _ConfigurationClientWrapper(_ConfigurationClientWrapperBase):
         feature_flag_selectors: List[SettingSelector],
         headers: Dict[str, str],
         **kwargs
-    ) -> Tuple[bool, Optional[Dict[Tuple[str, str], str]], Optional[List[Any]], Dict[str, bool]]:
+    ) -> Tuple[bool, Optional[Mapping[Tuple[str, str], Optional[str]]], Optional[List[Any]], Dict[str, bool]]:
         """
         Gets the refreshed feature flags if they have changed.
 
@@ -245,7 +250,7 @@ class _ConfigurationClientWrapper(_ConfigurationClientWrapperBase):
         return False, None, None, {}
 
     @distributed_trace
-    def get_configuration_setting(self, key: str, label: str, **kwargs) -> ConfigurationSetting:
+    def get_configuration_setting(self, key: str, label: str, **kwargs) -> Optional[ConfigurationSetting]:
         """
         Gets a configuration setting from the replica client.
 
@@ -291,6 +296,7 @@ class ConfigurationClientManager(ConfigurationClientManagerBase):  # pylint:disa
         replica_discovery_enabled,
         min_backoff_sec,
         max_backoff_sec,
+        load_balance: bool,
         **kwargs
     ):
         super(ConfigurationClientManager, self).__init__(
@@ -301,10 +307,13 @@ class ConfigurationClientManager(ConfigurationClientManagerBase):  # pylint:disa
             replica_discovery_enabled,
             min_backoff_sec,
             max_backoff_sec,
+            load_balance,
             **kwargs
         )
         self._original_connection_string = connection_string
         self._credential = credential
+        self._replica_clients: List[_ConfigurationClientWrapper] = []
+        self._active_clients: List[_ConfigurationClientWrapper] = []
 
         if connection_string and endpoint:
             self._replica_clients.append(
@@ -313,6 +322,8 @@ class ConfigurationClientManager(ConfigurationClientManagerBase):  # pylint:disa
                 )
             )
             self._setup_failover_endpoints()
+            if load_balance:
+                random.shuffle(self._replica_clients)
             return
         if endpoint and credential:
             self._replica_clients.append(
@@ -321,28 +332,80 @@ class ConfigurationClientManager(ConfigurationClientManagerBase):  # pylint:disa
                 )
             )
             self._setup_failover_endpoints()
+            if load_balance:
+                random.shuffle(self._replica_clients)
             return
         raise ValueError("Please pass either endpoint and credential, or a connection string with a value.")
+
+    def get_next_client(self) -> Optional[_ConfigurationClientWrapper]:
+        """
+        Get the next client to be used for the request. find_active_clients needs be called before this method or None
+        client will be returned. If load_balance isn't enabled the first active client will be returned, which is the
+        provided endpoint unless it has failed. If load_balance is enabled the next client in the list will be returned.
+
+        :return: The next client to be used for the request.
+        """
+        if not self._active_clients:
+            return None
+        if not self._load_balance:
+            return self._active_clients[0]
+        next_client = self._active_clients.pop(0)
+        self._last_active_client_name = next_client.endpoint
+        return next_client
+
+    def has_next_client(self) -> bool:
+        """
+        Check if there is another client to be used for the request. find_active_clients needs be called before this
+        method or False will be returned.
+
+        :return: True if there is another client to be used for the request, False otherwise.
+        :rtype: bool
+        """
+        return bool(self._active_clients)
+
+    def find_active_clients(self):
+        """
+        Figures out the current active clients, if load_balance is enabled the start of the list will be the one after
+        the last active client used.
+        """
+        active_clients = [client for client in self._replica_clients if client.is_active()]
+
+        self._active_clients = active_clients
+        if not self._load_balance or len(self._last_active_client_name) == 0:
+            return
+        for i, client in enumerate(active_clients):
+            if client.endpoint == self._last_active_client_name:
+                swap_point = (i + 1) % len(active_clients)
+                self._active_clients = active_clients[swap_point:] + active_clients[:swap_point]
+                return
 
     def refresh_clients(self):
         if not self._replica_discovery_enabled:
             return
         if self._next_update_time > time.time():
             return
-        self._setup_failover_endpoints()
+        updated_endpoints = self._setup_failover_endpoints()
+        if updated_endpoints and self._load_balance:
+            # Reshuffle only if a new client was added and load_balance is enabled
+            random.shuffle(self._replica_clients)
 
-    def _setup_failover_endpoints(self):
+    def get_client_count(self) -> int:
+        return len(self._replica_clients)
+
+    def _setup_failover_endpoints(self) -> bool:
         failover_endpoints = find_auto_failover_endpoints(self._original_endpoint, self._replica_discovery_enabled)
 
         if failover_endpoints is None:
             # SRV record not found, so we should refresh after a longer interval
             self._next_update_time = time.time() + FALLBACK_CLIENT_REFRESH_EXPIRED_INTERVAL
-            return
+            return False
 
         if len(failover_endpoints) == 0:
             # No failover endpoints in SRV record.
             self._next_update_time = time.time() + MINIMAL_CLIENT_REFRESH_INTERVAL
-            return
+            return False
+
+        updated_endpoints = False
 
         new_clients = [self._replica_clients[0]]  # Keep the original client
         for failover_endpoint in failover_endpoints:
@@ -353,6 +416,7 @@ class ConfigurationClientManager(ConfigurationClientManagerBase):  # pylint:disa
                     found_client = True
                     break
             if not found_client:
+                updated_endpoints = True
                 if self._original_connection_string:
                     failover_connection_string = self._original_connection_string.replace(
                         self._original_endpoint, failover_endpoint
@@ -367,7 +431,7 @@ class ConfigurationClientManager(ConfigurationClientManagerBase):  # pylint:disa
                             **self._args
                         )
                     )
-                else:
+                elif self._credential:
                     new_clients.append(
                         _ConfigurationClientWrapper.from_credential(
                             failover_endpoint,
@@ -380,6 +444,7 @@ class ConfigurationClientManager(ConfigurationClientManagerBase):  # pylint:disa
                     )
         self._replica_clients = new_clients
         self._next_update_time = time.time() + MINIMAL_CLIENT_REFRESH_INTERVAL
+        return updated_endpoints
 
     def backoff(self, client: _ConfigurationClientWrapper):
         client.failed_attempts += 1